Refactor: Split main.py into variant_validator and clinvar modules with docstrings
